requirements_dev.txt
imageio==2.9.0
importlib-metadata==3.3.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
matplotlib==3.3.3
notebook==6.1.6
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.0.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
scikit-learn==0.24.2
tqdm==4.60.0
seaborn==0.11.1